feature: resolved bug in goddard rocket

+++ src/environments/goddard.py
@@ -77,14 +77,14 @@ class CustomGoddardEnv(gym.Env):
             low   = np.array([0.0]),
             high  = np.array([1.0]),
             shape = (1,),
-            dtype = np.float
+            dtype = float
         )
 
         self.V_INDEX, self.H_INDEX, self.M_INDEX = 0, 1, 2
         self.observation_space = gym.spaces.Box(
-            low   = np.array([np.finfo(np.float).min, 0.0, self._r.M1]),
-            high  = np.array([np.finfo(np.float).max, np.finfo(np.float).max, self._r.M0]),
-            dtype = np.float
+            low   = np.array([np.finfo(float).min, 0.0, self._r.M1]),
+            high  = np.array([np.finfo(float).max, np.finfo(float).max, self._r.M0]),
+            dtype = float
         )
 
         self.reset()
